feat(ui): move useAntdTrableProps to ui
